const-eval: add `Const` and `Scalar` printing utilities (from `ir-utils`)
