+++ compiler/hash-ir-utils/Cargo.toml
@@ -11,6 +11,7 @@ doctest = false
 html-escape = "0.2.12"
 
 hash-ast = { path = "../hash-ast" }
+hash-const-eval = { path = "../hash-const-eval" }
 hash-ir = { path = "../hash-ir" }
 hash-layout = { path = "../hash-layout" }
 hash-source = { path = "../hash-source" }
