+++ compiler/hash-ir-utils/src/graphviz.rs
@@ -9,13 +9,14 @@
 
 use std::io;
 
+use hash_const_eval::print::pretty_print_const;
 use hash_ir::ir::{BasicBlock, BasicBlockData, Body, BodySource, TerminatorKind};
 use hash_layout::{compute::LayoutComputer, constant::Const};
 use hash_target::data_layout::HasDataLayout;
 use hash_utils::derive_more::Constructor;
 use html_escape::encode_text;
 
-use crate::{pretty_print_const, WriteIr};
+use crate::WriteIr;
 
 /// Used to separate line statements between each declaration within
 /// a particular body graph.
